Split out parser process lists into their own files
